requirement.txt
Package             Version
------------------- -------
astroid             2.2.5  
Django              2.2.1  
django-leaflet      0.24.0 
djangorestframework 3.9.4  
isort               4.3.20 
lazy-object-proxy   1.4.1  
mccabe              0.6.1  
pip                 19.1.1 
psycopg2            2.8.2  
psycopg2-binary     2.8.2  
pylint              2.3.1  
pytz                2019.1 
setuptools          41.0.1 
six                 1.12.0 
sqlparse            0.3.0  
typed-ast           1.3.5  
wheel               0.33.4 
wrapt               1.11.1 
